ci(release): qualify the atomic splash-to-shell handoff
