+++ tests/qualification/launcher/historical_launch/test_contract.py
@@ -87,11 +87,18 @@ def test_historical_shell_requires_ordered_trace_and_live_handoff_process(
     trace_path.parent.mkdir(parents=True)
     trace_path.write_text(
         "".join(
-            json.dumps({"event": event}) + "\n"
-            for event in (
-                "launch_splash.started",
-                "launch_splash.closed",
-                "main_shell.shown",
+            json.dumps(payload) + "\n"
+            for payload in (
+                {"event": "launch_splash.started"},
+                {"event": "main_shell.shown"},
+                {
+                    "event": "startup.visibility.first_event",
+                    "fields": {
+                        "event_type": "Paint",
+                        "label": "shell_frame",
+                    },
+                },
+                {"event": "launch_splash.closed"},
             )
         ),
         encoding="utf-8",
@@ -135,11 +142,18 @@ def test_historical_shell_rejects_trace_without_live_app_owner(
     trace_path.parent.mkdir(parents=True)
     trace_path.write_text(
         "".join(
-            json.dumps({"event": event}) + "\n"
-            for event in (
-                "launch_splash.started",
-                "launch_splash.closed",
-                "main_shell.shown",
+            json.dumps(payload) + "\n"
+            for payload in (
+                {"event": "launch_splash.started"},
+                {"event": "main_shell.shown"},
+                {
+                    "event": "startup.visibility.first_event",
+                    "fields": {
+                        "event_type": "Paint",
+                        "label": "shell_frame",
+                    },
+                },
+                {"event": "launch_splash.closed"},
             )
         ),
         encoding="utf-8",
